backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
supabase>=2.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
openai>=1.0.0
streamlit>=1.30.0
pandas>=2.0.0
